# src/main/java/com/bank/islamic/financialcrimemonitoring/api/ServiceDomainController.java
package com.bank.islamic.financialcrimemonitoring.api;

import com.bank.islamic.financialcrimemonitoring.model.ControlRecord;
import com.bank.islamic.financialcrimemonitoring.service.ControlRecordStore;
import io.github.resilience4j.circuitbreaker.annotation.CircuitBreaker;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.*;

import java.util.Collection;
import java.util.Map;

/**
 * BIAN semantic API for the "Financial Crime Monitoring" service domain.
 *
 * Endpoints follow the BIAN action-term style:
 *   GET  /v1/service-domain                          → who am I (SD metadata)
 *   POST /v1/financial-crime-alert-monitoring-state/initiate                    → Initiate a control record
 *   GET  /v1/financial-crime-alert-monitoring-state                             → Retrieve (list)
 *   GET  /v1/financial-crime-alert-monitoring-state/{crId}/retrieve             → Retrieve (single)
 *   PUT  /v1/financial-crime-alert-monitoring-state/{crId}/update               → Update
 *   PUT  /v1/financial-crime-alert-monitoring-state/{crId}/control              → Control (suspend|resume|terminate)
 */
@RestController
@RequestMapping("/v1")
public class ServiceDomainController {

    private final ControlRecordStore store;

    public ServiceDomainController(ControlRecordStore store) {
        this.store = store;
    }

    @GetMapping("/service-domain")
    public Map<String, String> serviceDomain() {
        return Map.of(
                "serviceDomain", "Financial Crime Monitoring",
                "businessArea", "Takaful and Shared Operations",
                "businessDomain", "Shared Operations",
                "functionalPattern", "Monitor",
                "assetType", "Financial Crime Alert",
                "controlRecord", "Financial Crime Alert Monitoring State",
                "version", "0.1.0",
                "phase", "1-shallow"
        );
    }

    @PostMapping("/financial-crime-alert-monitoring-state/initiate")
    @CircuitBreaker(name = "serviceDomain")
    public ResponseEntity<ControlRecord> initiate(@RequestBody(required = false) Map<String, Object> properties) {
        return ResponseEntity.status(HttpStatus.CREATED).body(store.initiate(properties));
    }

    @GetMapping("/financial-crime-alert-monitoring-state")
    public Collection<ControlRecord> list() {
        return store.list();
    }

    @GetMapping("/financial-crime-alert-monitoring-state/{crId}/retrieve")
    public ResponseEntity<ControlRecord> retrieve(@PathVariable String crId) {
        return store.retrieve(crId)
                .map(ResponseEntity::ok)
                .orElse(ResponseEntity.notFound().build());
    }

    @PutMapping("/financial-crime-alert-monitoring-state/{crId}/update")
    public ResponseEntity<ControlRecord> update(@PathVariable String crId,
                                                @RequestBody Map<String, Object> properties) {
        return store.update(crId, properties)
                .map(ResponseEntity::ok)
                .orElse(ResponseEntity.notFound().build());
    }

    @PutMapping("/financial-crime-alert-monitoring-state/{crId}/control")
    public ResponseEntity<?> control(@PathVariable String crId,
                                     @RequestBody Map<String, String> body) {
        try {
            return store.control(crId, body.get("action"))
                    .<ResponseEntity<?>>map(ResponseEntity::ok)
                    .orElse(ResponseEntity.notFound().build());
        } catch (IllegalArgumentException e) {
            return ResponseEntity.badRequest().body(Map.of("error", e.getMessage()));
        }
    }
}